Move get_related_fields() from cleanup app to core app
